Added AFDataReference
closes #528

Added new data_reference property
Fixed circular import errors
